♻️ Move components for ERDRenderer, including Spinner, TableNode, RelationshipEdge, and related utilities
