+++ .changeset/green-spies-itch.md
@@ -1,0 +1,5 @@
+---
+"@liam-hq/erd-core": patch
+---
+
+♻️ Move components for ERDRenderer, including Spinner, TableNode, RelationshipEdge, and related utilities
