+++ frontend/packages/erd-core/src/components/ERDRenderer/components/ERDContent/components/TableNode/TableDetail/CollapsibleHeader/CollapsibleHeader.tsx
