+++ frontend/packages/erd-core/src/components/ERDRenderer/components/ERDContent/components/TableNode/TableDetail/RelatedTables/RelatedTables.tsx
@@ -1,7 +1,7 @@
 import { GotoIcon, IconButton, Waypoints as WaypointsIcon } from '@liam-hq/ui'
 import { type Edge, type Node, ReactFlowProvider } from '@xyflow/react'
 import { type FC, type MouseEvent, useCallback } from 'react'
-import { ERDContent } from '../../../ERDContent'
+import { ERDContent } from '../../../../ERDContent'
 import { CollapsibleHeader } from '../CollapsibleHeader'
 import styles from './RelatedTables.module.css'
