+++ frontend/packages/erd-core/src/components/ERDRenderer/components/ERDContent/components/TableNode/index.ts
