+++ frontend/packages/erd-core/src/components/ERDRenderer/components/TableDetailDrawer/TableDetailDrawer.tsx
@@ -6,7 +6,7 @@ import {
 import { DrawerContent, DrawerPortal, DrawerRoot } from '@liam-hq/ui'
 import type { FC, PropsWithChildren } from 'react'
 // biome-ignore lint/nursery/useImportRestrictions: Fixed in the next PR.
-import { TableDetail } from '../ERDContent/TableNode/TableDetail'
+import { TableDetail } from '../ERDContent/components/TableNode/TableDetail'
 import styles from './TableDetailDrawer.module.css'
 
 const handleClose = () => updateActiveTableName(undefined)
